Time and dribs
