+++ dribdat/templates/admin/index.html
@@ -60,7 +60,7 @@
 
   <div class="mt-4">
     <a href="https://dribdat.cc" target="_blank"><img src="{{ url_for('static', filename='img/logo/logo13.png') }}" height="200" alt="dribdat logo"></a>
-    <br><br>
+    <p>{{ timeinfo }}</p>
     <a href="{{ url_for('public.clear_cache') }}">Refresh</a> |
     <b><a href="{{ url_for('public.about') }}">Documentation</a></b> |
     <a href="mailto:dribdat@datalets.ch">Get support</a>
