+++ dribdat/templates/public/dribs.html
@@ -64,11 +64,11 @@
           {% if s.ref_url %}
             </a>
           {% endif %}
-
+        </div>
+        
           {% if config.DRIBDAT_SOCIAL_LINKS and s.share %}
             {{ social.share_links(s.share.text, s.share.url) }}
           {% endif %}
-        </div>
       </div>
       {% endfor %}
     </section>
